+++ src/utility/DataMap.ts
@@ -3,7 +3,7 @@ import {
   SubjectsAndSchools,
   UniversitySchools,
 } from '@/types/BaseTypes';
-import { RawCourseData } from '@/types/PlanTypes';
+import { CourseTopicsData, RawCourseData } from '@/types/PlanTypes';
 import { ScheduleCourse } from '@/types/ScheduleTypes';
 import SchoolsJson from '@/data/schools.json';
 
@@ -39,24 +39,28 @@ export function subjectsAndSchools(data: any): SubjectsAndSchools {
 }
 
 export function plan({ courses, legacy, ...rest }: any): RawCourseData {
-  const courseMap = ({ i, n, u, r, d, p, s, f, c, t, o, l }: any) => ({
-    id: i,
-    name: n,
-    units: u,
-    repeatable: r,
-    description: d,
-    prereqs: p,
-    distros: s,
-    disciplines: f,
-    school: c,
-    terms: t,
-    topics: o,
-    placeholder: l,
-  });
+  const topics: CourseTopicsData = {};
+  const courseMap = ({ i, n, u, r, d, p, s, f, c, t, o, l }: any) => {
+    topics[i] = o;
+    return {
+      id: i,
+      name: n,
+      units: u,
+      repeatable: r,
+      description: d,
+      prereqs: p,
+      distros: s,
+      disciplines: f,
+      school: c,
+      terms: t,
+      placeholder: l,
+    };
+  };
 
   return {
     courses: courses.map(courseMap),
     legacy: legacy.map(courseMap),
+    topics,
     ...rest,
   };
 }